requirements.txt
# Core dependencies
python-telegram-bot==20.7
python-dotenv==1.1.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.10

# OpenAI and ML
openai==1.79.0
scikit-learn==1.5.0
numpy==2.2.6
pandas==2.2.3

# Data processing and parsing
pdfplumber==0.10.4
openpyxl==3.1.2
PyPDF2==3.0.1
xlrd==2.0.1
pytesseract==0.3.13

# Visualization
matplotlib==3.10.3
seaborn==0.13.2
pillow==11.2.1

# Utils
aiofiles==24.1.0
python-dateutil==2.9.0.post0
pytz==2025.2

# Optional: for better performance
gunicorn==21.2.0
